refactor(currencies): centralize zod schemas/types and dedupe test fixtures
Reorganize the api-currency-token package after review:
- schema.ts holds every Zod schema; types.ts holds all inferred + hand-written types.
- Move SYNC_HASH_QUERY_KEY to internals/constants and CalApiExtraSchema to schema.ts.
- Add src/fixtures (test-only) for mockApiTokenResponse/mockTokenCurrency and a buildApiTokenData
  factory, removing duplicated fixtures across api/utils/converter tests (Sonar duplication).
