+++ domain/api/currency-token/src/converter.test.ts
@@ -1,18 +1,16 @@
-import { convertApiToken, type ApiTokenData } from "./converter";
+import { convertApiToken } from "./converter";
+import { buildApiTokenData } from "./fixtures";
 
 describe("convertApiToken", () => {
   describe("Cardano transformation", () => {
     it("should not reconstruct if tokenIdentifier is missing", () => {
-      const apiToken: ApiTokenData = {
-        id: "cardano/native/policyId",
-        contractAddress: "policyId",
-        name: "Test Token",
-        ticker: "TEST",
-        units: [{ code: "TEST", name: "Test Token", magnitude: 6 }],
-        standard: "native",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(
+        buildApiTokenData({
+          id: "cardano/native/policyId",
+          contractAddress: "policyId",
+          standard: "native",
+        }),
+      );
 
       expect(result?.contractAddress).toBe("policyId");
     });
@@ -20,17 +18,14 @@ describe("convertApiToken", () => {
 
   describe("Sui passthrough", () => {
     it("should keep 'coin' standard as tokenType for sui tokens", () => {
-      const apiToken: ApiTokenData = {
-        id: "sui/coin/usdc_0xdba34672e30cb065b1f93e3ab55318768fd6fef66c15942c9f7cb846e2f900e7::usdc::usdc",
-        contractAddress:
-          "0xdba34672e30cb065b1f93e3ab55318768fd6fef66c15942c9f7cb846e2f900e7::usdc::USDC",
-        name: "USDC",
-        ticker: "USDC",
-        units: [{ code: "USDC", name: "USDC", magnitude: 6 }],
-        standard: "coin",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(
+        buildApiTokenData({
+          id: "sui/coin/usdc_0xdba34672e30cb065b1f93e3ab55318768fd6fef66c15942c9f7cb846e2f900e7::usdc::usdc",
+          contractAddress:
+            "0xdba34672e30cb065b1f93e3ab55318768fd6fef66c15942c9f7cb846e2f900e7::usdc::USDC",
+          standard: "coin",
+        }),
+      );
 
       expect(result?.tokenType).toBe("coin");
       expect(result?.parentCurrencyId).toBe("sui");
@@ -39,32 +34,13 @@ describe("convertApiToken", () => {
 
   describe("ledgerSignature handling", () => {
     it("should include ledgerSignature when provided", () => {
-      const apiToken: ApiTokenData = {
-        id: "ethereum/erc20/usdc",
-        contractAddress: "0xA0b86",
-        name: "USD Coin",
-        ticker: "USDC",
-        units: [{ code: "USDC", name: "USD Coin", magnitude: 6 }],
-        standard: "erc20",
-        ledgerSignature: "3045022100...",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(buildApiTokenData({ ledgerSignature: "3045022100..." }));
 
       expect(result?.ledgerSignature).toBe("3045022100...");
     });
 
     it("should not include ledgerSignature when not provided", () => {
-      const apiToken: ApiTokenData = {
-        id: "ethereum/erc20/usdc",
-        contractAddress: "0xA0b86",
-        name: "USD Coin",
-        ticker: "USDC",
-        units: [{ code: "USDC", name: "USD Coin", magnitude: 6 }],
-        standard: "erc20",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(buildApiTokenData());
 
       expect(result?.ledgerSignature).toBeUndefined();
     });
@@ -72,47 +48,27 @@ describe("convertApiToken", () => {
 
   describe("disableCountervalue handling", () => {
     it("should set disableCountervalue for testnet currencies", () => {
-      const apiToken: ApiTokenData = {
-        id: "ethereum_sepolia/erc20/usdc",
-        contractAddress: "0x123",
-        name: "USD Coin",
-        ticker: "USDC",
-        units: [{ code: "USDC", name: "USD Coin", magnitude: 6 }],
-        standard: "erc20",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(
+        buildApiTokenData({ id: "ethereum_sepolia/erc20/usdc", contractAddress: "0x123" }),
+      );
 
       expect(result?.disableCountervalue).toBe(true);
     });
 
     it("should respect explicit disableCountervalue flag", () => {
-      const apiToken: ApiTokenData = {
-        id: "ethereum/erc20/test",
-        contractAddress: "0x123",
-        name: "Test Token",
-        ticker: "TEST",
-        units: [{ code: "TEST", name: "Test Token", magnitude: 18 }],
-        standard: "erc20",
-        disableCountervalue: true,
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(
+        buildApiTokenData({
+          id: "ethereum/erc20/test",
+          contractAddress: "0x123",
+          disableCountervalue: true,
+        }),
+      );
 
       expect(result?.disableCountervalue).toBe(true);
     });
 
     it("should not disable countervalue for a mainnet currency without the flag", () => {
-      const apiToken: ApiTokenData = {
-        id: "ethereum/erc20/usdc",
-        contractAddress: "0xA0b86",
-        name: "USD Coin",
-        ticker: "USDC",
-        units: [{ code: "USDC", name: "USD Coin", magnitude: 6 }],
-        standard: "erc20",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(buildApiTokenData());
 
       expect(result?.disableCountervalue).toBe(false);
     });
@@ -120,47 +76,25 @@ describe("convertApiToken", () => {
 
   describe("Edge cases", () => {
     it("should return undefined for unknown parent currency", () => {
-      const apiToken: ApiTokenData = {
-        id: "unknowncurrency/erc20/test",
-        contractAddress: "0x123",
-        name: "Test",
-        ticker: "TEST",
-        units: [{ code: "TEST", name: "Test", magnitude: 18 }],
-        standard: "erc20",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(
+        buildApiTokenData({ id: "unknowncurrency/erc20/test", contractAddress: "0x123" }),
+      );
 
       expect(result).toBeUndefined();
     });
 
     it("should handle delisted tokens", () => {
-      const apiToken: ApiTokenData = {
-        id: "ethereum/erc20/old",
-        contractAddress: "0x123",
-        name: "Old Token",
-        ticker: "OLD",
-        units: [{ code: "OLD", name: "Old Token", magnitude: 18 }],
-        standard: "erc20",
-        delisted: true,
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(
+        buildApiTokenData({ id: "ethereum/erc20/old", contractAddress: "0x123", delisted: true }),
+      );
 
       expect(result?.delisted).toBe(true);
     });
 
     it("should handle empty units array", () => {
-      const apiToken: ApiTokenData = {
-        id: "ethereum/erc20/test",
-        contractAddress: "0x123",
-        name: "Test",
-        ticker: "TEST",
-        units: [],
-        standard: "erc20",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(
+        buildApiTokenData({ id: "ethereum/erc20/test", contractAddress: "0x123", units: [] }),
+      );
 
       expect(result?.units).toEqual([]);
     });
@@ -168,16 +102,7 @@ describe("convertApiToken", () => {
 
   describe("Standard token types", () => {
     it("should convert ERC20 token", () => {
-      const apiToken: ApiTokenData = {
-        id: "ethereum/erc20/usdc",
-        contractAddress: "0xA0b86",
-        name: "USD Coin",
-        ticker: "USDC",
-        units: [{ code: "USDC", name: "USD Coin", magnitude: 6 }],
-        standard: "erc20",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(buildApiTokenData());
 
       expect(result?.type).toBe("TokenCurrency");
       expect(result?.tokenType).toBe("erc20");
@@ -185,48 +110,39 @@ describe("convertApiToken", () => {
     });
 
     it("should convert SPL token", () => {
-      const apiToken: ApiTokenData = {
-        id: "solana/spl/usdc",
-        contractAddress: "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
-        name: "USD Coin",
-        ticker: "USDC",
-        units: [{ code: "USDC", name: "USD Coin", magnitude: 6 }],
-        standard: "spl",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(
+        buildApiTokenData({
+          id: "solana/spl/usdc",
+          contractAddress: "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
+          standard: "spl",
+        }),
+      );
 
       expect(result?.tokenType).toBe("spl");
       expect(result?.parentCurrencyId).toBe("solana");
     });
 
     it("should convert TRC20 token", () => {
-      const apiToken: ApiTokenData = {
-        id: "tron/trc20/usdt",
-        contractAddress: "TR7NHqjeKQxGTCi8q8ZY4pL8otSzgjLj6t",
-        name: "Tether USD",
-        ticker: "USDT",
-        units: [{ code: "USDT", name: "Tether USD", magnitude: 6 }],
-        standard: "trc20",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(
+        buildApiTokenData({
+          id: "tron/trc20/usdt",
+          contractAddress: "TR7NHqjeKQxGTCi8q8ZY4pL8otSzgjLj6t",
+          standard: "trc20",
+        }),
+      );
 
       expect(result?.tokenType).toBe("trc20");
       expect(result?.parentCurrencyId).toBe("tron");
     });
 
     it("should convert ASA token", () => {
-      const apiToken: ApiTokenData = {
-        id: "algorand/asa/31566704",
-        contractAddress: "31566704",
-        name: "USDC",
-        ticker: "USDC",
-        units: [{ code: "USDC", name: "USDC", magnitude: 6 }],
-        standard: "asa",
-      };
-
-      const result = convertApiToken(apiToken);
+      const result = convertApiToken(
+        buildApiTokenData({
+          id: "algorand/asa/31566704",
+          contractAddress: "31566704",
+          standard: "asa",
+        }),
+      );
 
       expect(result?.tokenType).toBe("asa");
       expect(result?.parentCurrencyId).toBe("algorand");
