+++ domain/api/currency-token/src/internals/utils.test.ts
@@ -1,5 +1,4 @@
 import type { FetchBaseQueryMeta } from "@reduxjs/toolkit/query/react";
-import type { TokenCurrency } from "@domain/entity-currency-token";
 
 jest.mock("../converter", () => ({
   convertApiToken: jest.fn(),
@@ -12,34 +11,10 @@ import {
   transformApiTokenToTokenCurrency,
   validateAndTransformSingleTokenResponse,
 } from "./utils";
-import { ApiTokenResponseSchema, type ApiTokenResponse } from "../schema";
+import { ApiTokenResponseSchema } from "../schema";
+import { mockApiTokenResponse, mockTokenCurrency } from "../fixtures";
 
 const mockConvert = convertApiToken as jest.MockedFunction<typeof convertApiToken>;
-
-const mockApiTokenResponse: ApiTokenResponse = {
-  id: "ethereum/erc20/usd_coin",
-  contract_address: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
-  name: "USD Coin",
-  ticker: "USDC",
-  units: [{ code: "USDC", name: "USD Coin", magnitude: 6 }],
-  standard: "erc20",
-  decimals: 6,
-  delisted: false,
-  live_signature: "3045022100...",
-};
-
-const mockTokenCurrency = {
-  type: "TokenCurrency",
-  id: "ethereum/erc20/usd_coin",
-  contractAddress: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
-  parentCurrencyId: "ethereum",
-  tokenType: "erc20",
-  name: "USD Coin",
-  ticker: "USDC",
-  delisted: false,
-  disableCountervalue: false,
-  units: [{ code: "USDC", name: "USD Coin", magnitude: 6 }],
-} as TokenCurrency;
 
 beforeEach(() => {
   jest.clearAllMocks();
