+++ domain/api/currency-token/src/types.ts
@@ -1,4 +1,11 @@
+import { z } from "zod";
 import type { TokenCurrency } from "@domain/entity-currency-token";
+import {
+  ApiTokenResponseSchema,
+  CalApiExtraSchema,
+  PersistedCALSchema,
+  PersistedTokenEntrySchema,
+} from "./schema";
 
 export interface TokenByIdParams {
   id: string;
@@ -46,3 +53,15 @@ export interface TokenPagination {
   /** Cursor for the next page of results. */
   nextCursor?: string;
 }
+
+/** A single token in the CAL `/v1/tokens` response (inferred from {@link ApiTokenResponseSchema}). */
+export type ApiTokenResponse = z.infer<typeof ApiTokenResponseSchema>;
+
+/** A persisted token entry: a {@link TokenCurrency} plus cache-restoration metadata. */
+export type PersistedTokenEntry = z.infer<typeof PersistedTokenEntrySchema>;
+
+/** The root persisted CAL blob (version pin + token entries + optional hash map). */
+export type PersistedCAL = z.infer<typeof PersistedCALSchema>;
+
+/** Thunk `extraArgument` contract for the CAL token api. */
+export type CalApiExtra = z.infer<typeof CalApiExtraSchema>;
